Improve javadoc for fluid interface generators and break method/field proxy generator in two.
