+++ artemis-fluid/artemis-fluid-core/src/main/java/com/artemis/generator/strategy/e/ComponentAccessorStrategy.java
@@ -8,7 +8,7 @@ import com.artemis.generator.util.MethodBuilder;
 import com.artemis.generator.util.Strings;
 
 /**
- * Generate direct accessor for each component.
+ * Adds method to get component from entity.
  *
  * @author Daan van Yperen
  */
