+++ artemis-fluid/artemis-fluid-core/src/main/java/com/artemis/generator/strategy/e/ComponentFieldAccessorStrategy.java
@@ -15,6 +15,11 @@ import java.util.Set;
 import static org.reflections.ReflectionUtils.*;
 
 /**
+ * Adds methods to interact with component fields.
+ *
+ * {@code class Pos extends Component { int x, y };} would extend the fluid interface with things
+ * like {@code e.posX(50).posY(20); float x = e.posX(); }
+ *
  * @author Daan van Yperen
  */
 public class ComponentFieldAccessorStrategy extends IterativeModelStrategy {
@@ -23,15 +28,6 @@ public class ComponentFieldAccessorStrategy extends IterativeModelStrategy {
     protected void apply(ComponentDescriptor component, TypeModel model) {
         final Set<Field> fields = component.getAllPublicFields();
         exposeFields(component, model, fields);
-
-        final Set<Method> methods = component.getAllPublicMethods();
-        exposeMethods(component, model, methods);
-    }
-
-    private void exposeMethods(ComponentDescriptor component, TypeModel model, Set<Method> methods) {
-        for (Method method : methods) {
-            exposeOnFluidInterface(component, method, model);
-        }
     }
 
     private void exposeFields(ComponentDescriptor component, TypeModel model, Set<Field> fields) {
@@ -40,72 +36,18 @@ public class ComponentFieldAccessorStrategy extends IterativeModelStrategy {
         }
     }
 
-    private void exposeOnFluidInterface(ComponentDescriptor component, Method method, TypeModel model) {
-
-        if (isSetter(method)) {
-            model.add(methodSetterMethod(component, method));
-        } else if (component.getPreferences().swallowGettersWithParameters && isGetterWithParameters(method)) {
-            // by preference, call getters and swallow the returned value, instead returning the fluid interface.
-            model.add(methodSetterMethod(component, method));
-        } else
-            // return the return type.
-            model.add(methodGetterMethod(component, method));
-    }
-
-    private boolean isGetterWithParameters(Method method) {
-        return !isSetter(method) && method.getParameterTypes().length > 0;
-    }
-
-    private boolean isSetter(Method method) {
-        return void.class.equals(method.getReturnType());
-    }
-
-
     private void exposeOnFluidInterface(ComponentDescriptor component, Field field, TypeModel model) {
-        if (0 == (Modifier.FINAL & field.getModifiers()))
-	        model.add(fieldSetterMethod(component, field));
+
+        // do not expose setters on final fields.
+        if (0 == (Modifier.FINAL & field.getModifiers())) {
+            model.add(fieldSetterMethod(component, field));
+        }
 
 	    model.add(fieldGetterMethod(component, field));
     }
 
-
-    private MethodDescriptor methodSetterMethod(ComponentDescriptor component, Method method) {
-        MethodBuilder builder = new MethodBuilder(FluidTypes.E_TYPE, component.getCompositeName(method.getName()));
-
-        String arguments = appendParameters(method, builder);
-
-        return builder
-                .mapper(component, ".create(entityId)." + method.getName() + "(" + arguments + ")")
-                .debugNotes(method.toGenericString())
-                .returnFluid()
-                .build();
-    }
-
-    private String appendParameters(Method method, MethodBuilder builder) {
-        int count = 0;
-        String arguments = "";
-        for (Class<?> parameterType : method.getParameterTypes()) {
-            builder.parameter(parameterType, "p" + count);
-            arguments = arguments + (arguments.isEmpty() ? "" : ", ") + "p" + count;
-            count++;
-        }
-        return arguments;
-    }
-
-    private MethodDescriptor methodGetterMethod(ComponentDescriptor component, Method method) {
-
-        MethodBuilder builder = new MethodBuilder(method.getGenericReturnType(), component.getCompositeName(method.getName()));
-
-        String arguments = appendParameters(method, builder);
-
-        return builder
-                .mapper("return ", component, ".create(entityId)." + method.getName() + "(" + arguments + ")")
-                .debugNotes(method.toGenericString())
-                .build();
-    }
-
     /**
-     * T componentName() -> create new entity.
+     * int E::posX() -> obtain field directly via interface.
      */
     private MethodDescriptor fieldGetterMethod(ComponentDescriptor component, Field field) {
         return new MethodBuilder(field.getGenericType(), component.getCompositeName(field.getName()))
@@ -115,7 +57,7 @@ public class ComponentFieldAccessorStrategy extends IterativeModelStrategy {
     }
 
     /**
-     * T componentName() -> create new entity.
+     * E E::posX(10) -> set field, returns fluid interface.
      */
     private MethodDescriptor fieldSetterMethod(ComponentDescriptor component, Field field) {
         final String parameterName = field.getName();
